molequle/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-dotenv
requests
pandas
numpy 